Update capture_apy.py
Change is within _send_request().
Commented urljoin line and instead joined it manually.
Using urljoin would result in passing a url that only includes the URI and results in a requests.exceptions.MissingSchema exception. The traceback and error text suggest the issue is the provided Capture API server address is missing. Only the URI is sent.
The new url string works.